requirements.txt
streamlit
torch
torchvision
transformers
pillow
plotly